requirements.txt
flask
sqlalchemy
sqlalchemy_utils
pymysql
paho-mqtt
